rename honeywell to quantinuum
