+++ mypy.ini
@@ -15,7 +15,7 @@ mypy_path =
     packages/openfermion,
     packages/stim,
     packages/openqasm,
-    packages/honeywell,
+    packages/quantinuum,
     packages/ionq,
     packages/itensor,
     packages/pyscf,
