+++ packages/quantinuum/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning", "setuptools"]
 build-backend = "poetry_dynamic_versioning.backend"
 
 [tool.poetry]
-name = "quri-parts-honeywell"
+name = "quri-parts-quantinuum"
 version = "0.0.0"
 description = "A plugin to use Honeywell with QIRI Parts"
 license = "Apache-2.0"
