+++ requirements-local.txt
@@ -12,7 +12,7 @@
 # -e ./packages/openfermion
 # -e ./packages/stim
 # -e ./packages/openqasm
-# -e ./packages/honeywell
+# -e ./packages/quantinuum
 # -e ./packages/ionq
 # -e ./packages/pyscf
 # -e ./packages/itensor
